// src/app/models/analisis-facial.ts
export interface AnalisisFacial {
    "id": number
    "patronFacial": number
    "perfil": number
    "asimetria": number
    "alturaFE": number
    "anchuraFE": number
    "perfilMaxilar": number
    "perfilMandibular": number
    "surcoLM": number
    "labiosEr":number
    "idDatosGenerales": number
}